Extract deployment configuration processing module
